Create, edit and delete track (READ DESCRIPTION)
Somehow dir-paginate changes the object composition which is being used
in the pagination, making it impossible to call the $resource methods.
It was necessary to pass the TrackId to the deleteTrack() instead of the
object at the current ng-repeat state. Hence, inside that function, it
was necessary to get the track based on the id from the web api.